Updated README.md and created shapes.go

+++ constants.go
@@ -2,8 +2,8 @@ package gogfx
 
 const (
 	// x-y borders
-	MAX_X int8 = 7 // max X coordinate for an object
-	MAX_Y int8 = 7 // max Y coordinate for an object
+	MAX_X int = 7 // max X coordinate for an object
+	MAX_Y int = 7 // max Y coordinate for an object
 	// ENUM: keyboard controls
 	CON_UP KeyboardControl = iota
 	CON_DOWN
